+++ src/test/java/com/camping/tests/support/fixture/KioskTestFixture.java
@@ -1,4 +1,4 @@
-package com.camping.tests.fixture;
+package com.camping.tests.support.fixture;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -6,7 +6,7 @@ import io.restassured.response.Response;
 import java.util.List;
 import java.util.Map;
 
-import static com.camping.tests.helper.ApiHelper.createExtractableResponse;
+import static com.camping.tests.support.helper.ApiHelper.createExtractableResponse;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class KioskTestFixture {
