+++ src/test/java/com/camping/tests/support/helper/DeleteStrategy.java
@@ -1,21 +1,18 @@
-package com.camping.tests.helper;
+package com.camping.tests.support.helper;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
 
-public class PatchStrategy implements HttpMethodStrategy {
+public class DeleteStrategy implements HttpMethodStrategy {
 
     @Override
     public <T> ExtractableResponse<Response> execute(RequestSpecification requestSpec, String url, T body) {
-        RequestSpecification given = RestAssured.given().spec(requestSpec);
-        if (body != null) {
-            given = given.body(body);
-        }
-
-        return given.when()
-                .patch(url)
+        return RestAssured.given()
+                .spec(requestSpec)
+                .when()
+                .delete(url)
                 .then()
                 .log().all()
                 .extract();
@@ -23,6 +20,6 @@ public class PatchStrategy implements HttpMethodStrategy {
 
     @Override
     public boolean supports(HttpMethod method) {
-        return method == HttpMethod.PATCH;
+        return method == HttpMethod.DELETE;
     }
 }
