+++ src/test/java/com/camping/tests/support/helper/HttpMethod.java
@@ -1,4 +1,4 @@
-package com.camping.tests.helper;
+package com.camping.tests.support.helper;
 
 public enum HttpMethod {
     GET, POST, PUT, PATCH, DELETE
