+++ src/test/java/com/camping/tests/support/helper/StepContext.java
@@ -1,4 +1,4 @@
-package com.camping.tests.helper;
+package com.camping.tests.support.helper;
 
 import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
